+++ source/MaterialXTest/XmlIo.cpp
@@ -21,6 +21,7 @@ TEST_CASE("Load content", "[xmlio]")
         "CustomNode.mtlx",
         "Looks.mtlx",
         "MaterialGraphs.mtlx",
+        "MultiOutput.mtlx",
         "PaintMaterials.mtlx",
         "PreShaderComposite.mtlx",
     };